test: Disable test code
테스트코드 전면 개편을 위해 비활성화 합니다.
